[Chore] BASE_URL를  env파일로 관리
